+++ menue/mainwindow.cpp
@@ -45,7 +45,7 @@ Mainwindow::Mainwindow()
     sprite->setResAnim(pBackground);
     sprite->setPosition(-1, -1);
     // background should be last to draw
-    sprite->setPriority(static_cast<short>(Mainapp::ZOrder::Background));
+    sprite->setPriority(static_cast<qint32>(Mainapp::ZOrder::Background));
     sprite->setScaleX(Settings::getWidth() / pBackground->getWidth());
     sprite->setScaleY(Settings::getHeight() / pBackground->getHeight());
