+++ menue/wikimenu.cpp
@@ -33,7 +33,7 @@ Wikimenu::Wikimenu()
     sprite->setResAnim(pBackground);
     sprite->setPosition(-1, -1);
     // background should be last to draw
-    sprite->setPriority(static_cast<short>(Mainapp::ZOrder::Background));
+    sprite->setPriority(static_cast<qint32>(Mainapp::ZOrder::Background));
     sprite->setScaleX(Settings::getWidth() / pBackground->getWidth());
     sprite->setScaleY(Settings::getHeight() / pBackground->getHeight());
 
@@ -141,7 +141,7 @@ void Wikimenu::search(bool onlyTag)
         textField->setWidth(pBox->getWidth() - 18);
         textField->setX(13);
         textField->setY(5);
-        pBox->setPriority(static_cast<short>(Mainapp::ZOrder::Objects));
+        pBox->setPriority(static_cast<qint32>(Mainapp::ZOrder::Objects));
         this->m_MainPanel->addItem(pBox);
         // add some event handling :)
         pBox->addEventListener(oxygine::TouchEvent::OVER, [ = ](oxygine::Event*)
