+++ objects/base/dropdownmenusprite.cpp
@@ -11,7 +11,7 @@ DropDownmenuSprite::DropDownmenuSprite(qint32 width, QVector<QString>& items, st
     Q_ASSERT(items.size() != 0);
     Mainapp* pApp = Mainapp::getInstance();
     this->moveToThread(pApp->getWorkerthread());
-    this->setPriority(static_cast<short>(Mainapp::ZOrder::Objects));
+    this->setPriority(static_cast<qint32>(Mainapp::ZOrder::Objects));
     this->setWidth(width);
     m_pClipActor->setY(5);
 
@@ -29,7 +29,7 @@ DropDownmenuSprite::DropDownmenuSprite(qint32 width, QStringList& items, std::fu
     Q_ASSERT(items.size() != 0);
     Mainapp* pApp = Mainapp::getInstance();
     this->moveToThread(pApp->getWorkerthread());
-    this->setPriority(static_cast<short>(Mainapp::ZOrder::Objects));
+    this->setPriority(static_cast<qint32>(Mainapp::ZOrder::Objects));
     this->setWidth(width);
     m_pClipActor->setY(5);
